requirements.txt
ipython==5.7.0
notebook
gitpython
teamcity-messages
coverage
wheel
bumpversion